move SideBarNav into its own folder, delete HeuristicsActivity page
